+++ Runtime/Services/Attributes/AutoRegisteredService.cs
@@ -1,0 +1,8 @@
+using CinderUtils.Services;
+using System;
+
+
+namespace CinderUtils.Services {
+    [AttributeUsage(AttributeTargets.Class)]
+    public class AutoRegisteredServiceAttribute : Attribute { }
+}