+++ Runtime/Services/MonoBehaviourService.cs
@@ -1,0 +1,15 @@
+using System;
+using UnityEngine;
+
+
+namespace CinderUtils.Services {
+    public class MonoBehaviourService : MonoBehaviour, IService {
+        protected virtual void Awake() {
+            DontDestroyOnLoad(gameObject);
+        }
+
+        protected virtual void Reset() { 
+            gameObject.name = GetType().Name;
+        }
+    }
+}