+++ CHANGELOG.md
@@ -137,7 +137,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [unreleased]: https://github.com/ash-dvlpr/cinder-utils/compare/v1.2.0...master
-[v1.2.0]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.2.1
+[v1.2.0]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.2.0
 [v1.1.2]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.1.2
 [v1.1.1]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.1.1
 [v1.1.0]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.1.0